requirements.txt
Django==4.2
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-decouple==2.1
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==21.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
phonenumbers==8.13.30
requests~=2.31.0
